+++ app/Models/ClientDirector.php
@@ -9,12 +9,13 @@ class ClientDirector extends Model
 {
     protected $fillable = [
         'client_id', 'name', 'role', 'appointed_on', 'resigned_on',
-        'dob_month', 'dob_year', 'nationality', 'occupation', 'country_of_residence',
+        'dob_month', 'dob_year', 'nationality', 'occupation', 'country_of_residence', 'sa_required',
     ];
 
     protected $casts = [
         'appointed_on' => 'date',
         'resigned_on'  => 'date',
+        'sa_required'  => 'boolean',
     ];
 
     public function client(): BelongsTo